requirements.txt
# Silver Tier Functional Assistant - Python Dependencies
# Install with: pip install -r requirements.txt

# Core Dependencies
watchdog>=3.0.0
schedule>=1.2.0
requests>=2.31.0

# Gmail API (for Gmail Watcher)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.0.0

# LinkedIn Automation (optional)
# Note: Requires Chrome/Chromium browser
selenium>=4.15.0
playwright>=1.40.0
webdriver-manager>=4.0.0

# Environment Variables
python-dotenv>=1.0.0

# Utilities
pathlib2>=2.3.7; python_version < "3.4"
